+++ src/mine/ui/controls/api/include/mine/ui/controls/ControlsAll.h
@@ -9,6 +9,7 @@
 #include <mine/ui/controls/ModuleTag.h>
 #include <mine/ui/controls/ContentPresenter.h>
 #include <mine/ui/controls/ContentControl.h>
+#include <mine/ui/controls/UserControl.h>
 #include <mine/ui/controls/TextBlock.h>
 #include <mine/ui/controls/Button.h>
 #include <mine/ui/controls/Border.h>